% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_images.R
\name{lsd_images}
\alias{lsd_images}
\title{Function to perform line segment detection on images of a build}
\usage{
lsd_images(start_here, end_here, directory)
}
\arguments{
\item{start_here}{integer number of first file in build.}

\item{end_here}{integer number of last file in build.}

\item{directory}{string directory images are held.}
}
\value{
res vector count of detected line segments per layer.
}
\description{
Function to perform line segment detection on images of a build
}
